// autogpt_platform/frontend/src/app/api/__generated__/models/aPIKeyCredentials.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { APIKeyCredentialsTitle } from "./aPIKeyCredentialsTitle";
import type { APIKeyCredentialsExpiresAt } from "./aPIKeyCredentialsExpiresAt";

export interface APIKeyCredentials {
  id?: string;
  provider: string;
  title?: APIKeyCredentialsTitle;
  type?: "api_key";
  api_key: string;
  /** Unix timestamp (seconds) indicating when the API key expires (if at all) */
  expires_at?: APIKeyCredentialsExpiresAt;
}